PlaceOrderApp/requirement.txt
asgiref==3.3.1
astroid==2.4.2
beautifulsoup4==4.9.3
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
console-log==0.2.10
Django==3.1.3
gevent==20.9.0
gevent-websocket==0.10.1
greenlet==0.4.17
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pyactiveresource==2.2.1
pycparser==2.20
pylint==2.6.0
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
ShopifyAPI==8.0.4
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
wsgigzip==0.1.4
zope.event==4.5.0
zope.interface==5.2.0
